requirements.txt
click==8.0.1
coverage==5.5
cx-Freeze==6.8.1
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==4.8.1
numpy==1.21.2
Pillow==8.3.2
zipp==3.6.0
PySimpleGUI~=4.55.1